Finished Issue 3
Animations are now working for walking, dashing, jumping, and idle
