+++ ProjectEroica/Source/ProjectEroica/AnimInstanceKisa.cpp
@@ -1,0 +1,6 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "AnimInstanceKisa.h"
+
+
+
