--- ProjectEroica/Source/ProjectEroica/TP_SideScroller/TP_SideScroller.h
@@ -1,5 +1,0 @@
-// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
