+++ ProjectEroica/Source/ProjectEroica/TP_SideScrollerGameMode.cpp
@@ -1,0 +1,16 @@
+// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
+
+#include "TP_SideScrollerGameMode.h"
+#include "TP_SideScrollerCharacter.h"
+#include "UObject/ConstructorHelpers.h"
+
+ATP_SideScrollerGameMode::ATP_SideScrollerGameMode()
+{
+	// set default pawn class to our Blueprinted character
+	static ConstructorHelpers::FClassFinder<APawn> PlayerPawnBPClass(TEXT("/Game/SideScrollerCPP/Blueprints/SideScrollerCharacter"));
+	if (PlayerPawnBPClass.Class != NULL)
+	{
+		DefaultPawnClass = PlayerPawnBPClass.Class;
+	}
+	
+}
