+++ ProjectEroica/Source/ProjectEroica/TP_SideScrollerGameMode.h
@@ -1,0 +1,19 @@
+// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "TP_SideScrollerGameMode.generated.h"
+
+UCLASS(minimalapi)
+class ATP_SideScrollerGameMode : public AGameModeBase
+{
+	GENERATED_BODY()
+
+public:
+	ATP_SideScrollerGameMode();
+};
+
+
+
